chore(plugin-google-analytics): reestructure files
