+++ packages/botonic-plugin-google-analytics-v4/package.json
@@ -20,6 +20,7 @@
     "uuid": "^8.2.0"
   },
   "devDependencies": {
+    "@botonic/core": "^0.21.0",
     "@types/node": "^18.14.1",
     "typescript": "^4.9.5"
   },
