+++ packages/botonic-plugin-google-analytics-v4/src/index.ts
@@ -1,48 +1,35 @@
-export function getEnv(name: string): string {
-  const value = process.env[name]
-  if (value === undefined || value === null) {
-    throw 'Missing env var for ' + name
-  }
-  return value
-}
-
+import { PluginPostRequest, PluginPreRequest } from '@botonic/core'
 import axios from 'axios'
 
-export type GA4Options = {
-  clientId: string
-}
+import { BotSession, GA4Event, GA4Options } from './types'
 
-export type GA4Event = {
-  name: string
-  params?: GA4Params
-}
-
-export type GA4Params = Record<string, string | number>
-
-export type EventBodyParams = {
-  clientId: string
-  events: GA4Event[]
-}
+const defaultGetClientId = (session: BotSession) => session.user.id
 
 export default class BotonicPluginGoogleAnalytics4 {
   private baseUrl = 'https://www.google-analytics.com/mp/collect'
   private apiSecret: string
   private measurementId: string
-  private clientId: string
+  private getClientId: (session: BotSession) => string
   private userId?: string
   constructor(options: GA4Options) {
     this.apiSecret = getEnv('ga4ApiSecret')
     this.measurementId = getEnv('ga4MeasurementId')
-    this.clientId = options.clientId
+    this.getClientId = options.getClientId || defaultGetClientId
   }
 
-  public async track(events: GA4Event | GA4Event[]): Promise<void> {
+  pre(_request: PluginPreRequest): void {}
+  post(_request: PluginPostRequest): void {}
+
+  public async track(
+    session: BotSession,
+    events: GA4Event | GA4Event[]
+  ): Promise<void> {
     events = Array.isArray(events) ? events : [events]
 
     await axios.post(
       this.baseUrl,
       {
-        client_id: this.clientId,
+        client_id: this.getClientId(session),
         user_id: this.userId,
         events,
       },
@@ -54,4 +41,12 @@ export default class BotonicPluginGoogleAnalytics4 {
       }
     )
   }
+}
+
+export function getEnv(name: string): string {
+  const value = process.env[name]
+  if (value === undefined || value === null) {
+    throw 'Missing env var for ' + name
+  }
+  return value
 }
